+++ HISTORY.md
@@ -13,6 +13,7 @@
 * blocks, objects, threads: made AND variadic
 * blocks, objects, threads: made OR variadic
 * adjusted English translation: removed "any" -> "random" rewording
+* blocks: renamed all 'any' selectors in dropdowns into 'random'
 
 ### 2023-02-13
 * new dev version